Fixed many of the compile warnings. Added a "Start Servers.bat" to easily run the servers. Added an all-in-one project under ./win32/DS-server.sln Added the preprocessor definition of _CRT_SECURE_NO_WARNINGS to all projects. Expanded the WELL512 class, and commented the irand function as depreciated. The way people are using irand with % is very bad practice and shouldn't be used, and irand should be made private in the future. People need to stop using the irand function in code, and use the new functions in the WELL512 class: int32 GetRandomInt32(int32 min, int32 max) int32 GetRandomInt32(int32 max) int16 GetRandomInt16(int16 min, int16 max) int16 GetRandomInt16(int16 max) int8 GetRandomInt8(int8 min, int8 max) int8 GetRandomInt8(int8 max) float GetRandomFloat(float min, float max, bool allow_decimals = false) float GetRandomFloat(float max, bool allow_decimals = false)

+++ src/map/utils/puppetutils.cpp
@@ -244,7 +244,7 @@ void setAttachment(CCharEntity* PChar, uint8 slotId, uint8 attachment)
             {
                 for (int i = 0; i < 8; i++)
                 {
-                    PChar->PAutomaton->addElementCapacity(i, -((PAttachment->getElementSlots() >> (i * 4)) & 0xF));
+                    PChar->PAutomaton->addElementCapacity(i, -(int8)((PAttachment->getElementSlots() >> (i * 4)) & 0xF));
                     PChar->PAutomaton->setAttachment(slotId, 0);
                 }
             }
